requirements.txt
flask 
numpy 
pandas 
matplotlib 
pillow 
keras 
tensorflow